SF: Scheduler: opportunistically enter idle after touch
This commit drastically changes how the scheduler works.

Now, the device will be forced to maintain idle state (set to 60 Hz in
this commit) after touchboost is finished and idle state is entered once.

The main principle of this commit is that the highest refresh rate isn't
necessary for everything and applying it only for touch-related events
are plenty enough.

It's been deliberately set to 60 Hz so that anomalies from the nature of
imperfectiveness of this method is minimum and less-noticeable compared
to the lowest possible refresh rate.

Example of this commit in action:

<Before 1>
| Touch starts | Scrolling starts | Touch ends | Scrolling ends | Scroll-bar timeout | Scroll-bar fade-out animation | Idle  |
| 120 Hz       | 120 Hz           | 120 Hz     | 120 Hz         | 24 Hz              | 120 Hz                        | 24 Hz |

<After 1>
| Touch starts | Scrolling starts | Touch ends | Scrolling ends | Scroll-bar timeout | Scroll-bar fade-out animation | Idle  |
| 120 Hz       | 120 Hz           | 120 Hz     | 120 Hz         | 24 Hz              | 60 Hz                         | 24 Hz |

<Before 2>
| Idle  | Wi-Fi status bar icon changes | Timeout | Idle  |
| 24 Hz | 120 Hz                        | 120 Hz  | 24 Hz |

<After 2>
| Idle  | Wi-Fi status bar icon changes | Timeout | Idle  |
| 24 Hz | 60 Hz                         | 60 Hz   | 24 Hz |

[Jyotiraditya Panda <jyotiraditya@aospa.co>]
  - Adapt to AOSPA/uvite (Android 14)

Change-Id: I27dc60f35821ee4a5726f67f9b538b394194d00c
Co-authored-by: Jyotiraditya Panda <jyotiraditya@aospa.co>
Signed-off-by: Jyotiraditya Panda <jyotiraditya@aospa.co>
Signed-off-by: Omkar Chandorkar <gotenksIN@aospa.co>
Signed-off-by: Dmitrii <bankersenator@gmail.com>
Signed-off-by: Hưng Phan <phandinhhungvp2001@gmail.com>